feat(breakout-rooms): unify components into BreakoutRoom
