+++ react/features/breakout-rooms/components/native/BreakoutRooms.tsx
@@ -16,7 +16,7 @@ import {
 
 import AutoAssignButton from './AutoAssignButton';
 import BreakoutRoomsFooter from './BreakoutRoomsFooter';
-import { CollapsibleRoom } from './CollapsibleRoom';
+import { BreakoutRoom } from './BreakoutRoom';
 import styles from './styles';
 
 
@@ -38,10 +38,9 @@ const BreakoutRooms = () => {
             { showAutoAssign && <AutoAssignButton /> }
             {
                 isBreakoutRoomsSupported
-                && rooms.map(room => (<CollapsibleRoom
+                && rooms.map(room => (<BreakoutRoom
                     key = { room.id }
-                    room = { room }
-                    roomId = { room.id } />))
+                    room = { room } />))
             }
         </>
     ), [ showAutoAssign, isBreakoutRoomsSupported, rooms ]);
